Rename TinyDB to NameDB and simplify database operations for player name lookups
