+++ src/main/resources/config.yml
@@ -32,6 +32,7 @@ world:
   newbiekit:
   - "DIAMOND_PICKAXE:1"
   - "BREAD:2"
+  - "OAK_BOAT:1"
 
 # Lobby settings
 lobby:
